% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{aggregate_cells}
\alias{aggregate_cells}
\title{A wrapper around scuttle::aggregateAcrossCells that also:
\itemize{
\item computes normcounts (divide by library size) and logcounts (normcounts and take log with pseudocount 1)
\item adds QC metrics
\item marks mito genes
\item formats sample names
}}
\usage{
aggregate_cells(sce, criteria, assay = c("normcounts", "counts"))
}
\description{
A wrapper around scuttle::aggregateAcrossCells that also:
\itemize{
\item computes normcounts (divide by library size) and logcounts (normcounts and take log with pseudocount 1)
\item adds QC metrics
\item marks mito genes
\item formats sample names
}
}
