% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotate_mitochondrial_genes}
\alias{annotate_mitochondrial_genes}
\title{Annotate mitochondrial genes}
\usage{
annotate_mitochondrial_genes(sce, pattern = "mt:")
}
\arguments{
\item{pattern}{Character which if contained in a gene name signals it is a mitochondrial gene}
}
\description{
Set the mitochondrial field of the gene metadata to true
if \code{pattern} is found in the gene name
}
\details{
The mitochondria has its own genome and mitochondrial genes are encoded in this genome
i.e. genes encoded in the nucleus but whose protein product travels to the mitochondria are NOT mitochondrial genes
}
