% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{filter_sce}
\alias{filter_sce}
\title{Keep cells whose barcode is included in the barcodes_file}
\usage{
filter_sce(sce, barcodes_file, verbose = 0)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{barcodes_file}{A txt file leading to cell barcodes, one per line}
}
\description{
The cell barcode must be available
in the CellID column of the colData dataframe in the sce object
}
