% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{getDEGs}
\alias{getDEGs}
\title{Get DEGs found in a list of diff_tables}
\usage{
getDEGs(
  data_list,
  sign = 0,
  effect.size_threshold = 0.05,
  signif_threshold = 0.05
)
}
\description{
Get Differentially Expressed Genes (DEGs) across diff_table(s)
Optionally, customise the effect size and/or significance threshold
Optionally keep genes
}
