% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{keep_genes}
\alias{keep_genes}
\title{Filter sce so only genes are available
Sort the sce so genes are alphabetically ordered}
\usage{
keep_genes(sce, genes)
}
\description{
Filter sce so only genes are available
Sort the sce so genes are alphabetically ordered
}
