% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{label_cells}
\alias{label_cells}
\title{Label groups of cells given a comparison}
\usage{
label_cells(sce, comparison_file, remove = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{comparison_file}{Path to comparison file}

\item{remove}{If True, cells not included in the comparison
are removed from the SingleCellExperiment obhect}
}
\value{
Labeled (and optionally reduced) SingleCellExperiment object
}
\description{
Given a pairwise-comparison specifying cells,
an experiment_ column is added that specifies one of
c("Minus", "Plus", "")
}
