% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{loom2sce}
\alias{loom2sce}
\title{Load a Scope loomfile into R}
\usage{
loom2sce(path)
}
\arguments{
\item{path}{Path to a Scope loomfile}
}
\description{
Open a Scope loomfile and load its data
into R as a SingleCellExperiment object
}
\details{
A scope loomfile is an HDF5 file with the following structure:
\itemize{
\item \textbf{matrix}: counts matrix
\item \textbf{col_attrs}: cell metatadata. cell identifiers (barcodes) are available under CellID field
\item \textbf{row_attrs}: gene metatadata. gene identifiers are available under Gene field
\item Extra layers may be available
}
}
