% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{match_group}
\alias{match_group}
\title{Return index of cells in SingleCellExperiment that match group}
\usage{
match_group(sce, group)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{group}{named list matching column names in the sce colData}
}
\value{
Logical vector where ith value is TRUE if the ith cell belongs to group
}
\description{
Return index of cells in SingleCellExperiment that match group
}
\seealso{
\code{label_cells}
}
