% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot the histogram of counts of a gene across cells in a pairwise comparison}
\usage{
plotHistogram(
  sce,
  gene,
  comparison_file,
  by_exprs_values = "counts",
  geom = "histogram"
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{gene}{A gene contained in the Gene variable in the rowData of sce}

\item{comparison_file}{Path to a comparison file}

\item{by_exprs_values}{Name of a counts assay in the sce}
}
\description{
Plot the histogram of counts of a gene across cells in a pairwise comparison
}
\details{
\itemize{
\item The SingleCellObject must have a colData column called experiment_
encoding which group the corresponding cell belongs to
\item The colData should contain a CellID column unique to each cell
\item The SingleCellExperiment must have the assay specified by by_exprs_values i.e. at least counts
}
}
\seealso{
\code{readComparisonFile}
}
