% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{readComparisonFile}
\alias{readComparisonFile}
\title{Read a comparison file into R}
\usage{
readComparisonFile(path)
}
\arguments{
\item{path}{Path to a comparison file}
}
\description{
Read a comparison file into R
}
\details{
A comparison file is a csv file with a header whose column names
match column names in SingleCellExperiment objects
Each row specifies the set of cells matching all column values
If all possible values of a metavariable should be included in the group
i.e. the metavariable is to be ignored when building the group, supply * to match everything
A comparison file should also contain a column named experiment_ with values either Minus or Plus
Only pairwise comparisons are supported for now
}
