% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_io.R
\name{readEdgeR}
\alias{readEdgeR}
\title{Load diff_table produced by MAST}
\usage{
readEdgeR(path, signif_threshold = NULL, only_signif = FALSE, adjust = TRUE)
}
\description{
edgeR does not return adjusted PValues (for now)
but an integer called direction which is non 0 if the adj-PValue is significant
and 0 if it is not significant
If it is significant, it will be eithe 1 or -1 depending on the sign of the fold change
}
