% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_io.R
\name{readMAST}
\alias{readMAST}
\title{Load diff_table produced by MAST}
\usage{
readMAST(path, signif_threshold = 0.05, only_signif = FALSE)
}
\description{
Publication: https://genomebiology.biomedcentral.com/articles/10.1186/s13059-015-0844-5
Notebook: https://github.com/antortjim/SleepSignature/blob/main/notebooks/08-MAST.R.ipynb
}
