% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{sce2loom}
\alias{sce2loom}
\title{Export a SingleCellExperiment as a loomfile}
\usage{
sce2loom(sce, output_loomfile)
}
\arguments{
\item{sce}{SingleCellExperiment instance}

\item{output_loomfile}{Path to newly created loomfile. It must not exist@param clean If TRUE, all columns containing ClusterMarkers in gene metadata are dropped}
}
\description{
Export a SingleCellExperiment as a loomfile
}
