% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{snakemaker}
\alias{snakemaker}
\title{Load a json file and generate a fake Snakemake object
for easy debugging of Snakemake pipelines}
\usage{
snakemaker(jsonfile, overwrite = FALSE)
}
\description{
The snakemake object is bound to the global environment
using rlang functionality
}
